Исправил все ошибки и сделал немного следующее дз
