+++ js/data.js
@@ -1,6 +1,6 @@
-export const MIN = 1;
-export const MAX = 25;
-export const MAX_COMMENT_LENGTH = 140;
-export const VERIFIED_STRING = 'message';
-export const PICTURE_COUNT = 25;
+// export const MIN = 1;
+// export const MAX = 25;
+// export const MAX_COMMENT_LENGTH = 140;
+// export const VERIFIED_STRING = 'message';
+// export const PICTURE_COUNT = 25;
