+++ js/main.js
@@ -1,3 +1,4 @@
+import './send.js';
 import './download.js';
 import './form.js';
-import './send.js';
+// import './send.js';
