+++ js/util.js
@@ -1,26 +1,26 @@
 // Функция создания значений для фотографий
-const createPicture = (id) => ({
-  id: id,
-  url: `photos/${id}.jpg`,
-  description: 'Фото природы',
-  likes: getRandomNumber(15, 200),
-  comment: getRandomNumber(0, 200),
-});
+// const createPicture = (id) => ({
+//   id: id,
+//   url: `photos/${id}.jpg`,
+//   description: 'Фото природы',
+//   likes: getRandomNumber(15, 200),
+//   comment: getRandomNumber(0, 200),
+// });
 
-// Функция для проверки максимальной длины строки
-export function checkLengthString(verifiedString, maxLength) {
-  if (verifiedString.length <= maxLength) {
-    return true;
-  }
-  return false;
-}
+// // Функция для проверки максимальной длины строки
+// export function checkLengthString(verifiedString, maxLength) {
+//   if (verifiedString.length <= maxLength) {
+//     return true;
+//   }
+//   return false;
+// }
 
-// Функция, возвращающая случайное целое число из переданного диапазона включительно
-export function getRandomNumber (beginValue, endValue) {
-  if (beginValue >= 0 && endValue >= 0 && beginValue < endValue) {
-    return Math.round(Math.random() * (endValue - beginValue)) + beginValue;
-  }
-  return NaN;
-}
+// // Функция, возвращающая случайное целое число из переданного диапазона включительно
+// export function getRandomNumber (beginValue, endValue) {
+//   if (beginValue >= 0 && endValue >= 0 && beginValue < endValue) {
+//     return Math.round(Math.random() * (endValue - beginValue)) + beginValue;
+//   }
+//   return NaN;
+// }
 
-export {createPicture};
+// export {createPicture};
